requirements.txt
tweepy
httpx
py-getch
pillow
colorama
coloredlogs
backoff